+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 This project was forked from `py-solc`. View the original changelog [here](https://github.com/ethereum/py-solc/blob/master/CHANGELOG).
 
 ## [Unreleased](https://github.com/iamdefinitelyahuman/py-solc-x)
+- improve `solc` version detection and raise correct error if it fails
 
 ## [1.0.0](https://github.com/iamdefinitelyahuman/py-solc-x/releases/tag/v1.0.0) - 2020-08-26
 ### Added